Removed atexit callback
On linux this was causing a segfault when destroying the
SDL renderer. I couldn't figure out exactly why, but probably
something to do with when the handler is actually called by the OS
and then it automatically cleans up resources on exit. Im just manually
calling the handler before the app exits now. If there is a fatal error
initialising audio,video, etc then i'm not worrying about destroying
the resources.